backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
et_xmlfile==2.0.0
fastapi==0.128.0
gTTS==2.5.4
h11==0.16.0
holidays==0.88
idna==3.11
numpy==2.4.1
openpyxl==3.1.5
pandas==2.3.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-multipart==0.0.21
python-vlc==3.0.21203
pytz==2025.2
requests==2.32.5
schedule==1.2.2
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
yt-dlp==2025.12.8
edge-tts>=6.1.9
